requirements.txt
anthropic
gradio
python-dotenv
boto3